Crate futures_stop

An example implementation of allowing futures::Streams to be stopped remotely.

This is a very simple implementation which wraps a stream with another stream that will finish when it receives a message from a sender.

Examples

let (stream, handle) = futures_stop::convert_to_stoppable(stream);

// ...

handle.stop().wait().unwrap();
// or
handle.wait().stop();

Structs

Handle
A handle for a Stream which allows sending a stop message.

Stream
A stoppable version of the standard futures::Stream

Wait
A blocking, synchronous version of a Handle.

Functions

convert_to_stoppable
Converts a regular futures::Stream into a Stream and Handle pair.

Type Definitions

StopSender
Type alias for a Sender used for sending stop messages